Move NoLayerEmbedTokens
